requirements.txt
tqdm
openai
sentence-transformers
python-dotenv
scikit-learn
numpy
python-dateutil